+++ src/main/java/es/eoi/repository/ClasesReservasRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ClasesReservasRepository extends JpaRepository<ClasesReservas,Long> {
+public interface ClasesReservasRepository extends JpaRepository<ClasesReservas,Integer> {
 }
