+++ src/main/java/es/eoi/repository/DietistasRepository.java
@@ -3,5 +3,5 @@ package es.eoi.repository;
 import es.eoi.model.Dietistas;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DietistasRepository extends JpaRepository<Dietistas, Long> {
+public interface DietistasRepository extends JpaRepository<Dietistas, Integer> {
 }
