+++ src/main/java/es/eoi/repository/EntrenadoresRepository.java
@@ -4,5 +4,5 @@ package es.eoi.repository;
 import es.eoi.model.Entrenadores;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EntrenadoresRepository extends JpaRepository<Entrenadores, Long> {
+public interface EntrenadoresRepository extends JpaRepository<Entrenadores, Integer> {
 }
