+++ src/main/java/es/eoi/repository/PerfilUsuarioRepository.java
@@ -3,5 +3,5 @@ package es.eoi.repository;
 import es.eoi.model.PerfilUsuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PerfilUsuarioRepository extends JpaRepository<PerfilUsuario,Long> {
+public interface PerfilUsuarioRepository extends JpaRepository<PerfilUsuario,Integer> {
 }
