+++ src/main/java/es/eoi/repository/RecetasRespository.java
@@ -3,5 +3,5 @@ package es.eoi.repository;
 import es.eoi.model.Recetas;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RecetasRespository extends JpaRepository<Recetas, Long> {
+public interface RecetasRespository extends JpaRepository<Recetas, Integer> {
 }
